refactor: storage enums
