+++ src/components/player/PlayerModeButton.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import TrackPlayer, { RepeatMode } from 'react-native-track-player';
 import { IconButton } from 'react-native-paper';
 
-import { NoxRepeatMode } from './enums/RepeatMode';
+import { NoxRepeatMode } from '@enums/RepeatMode';
 import { useNoxSetting } from '@hooks/useSetting';
 import noxPlayingList from '@stores/playingList';
 import { savePlayMode } from '@utils/ChromeStorage';
