+++ src/components/player/usePlayerControls.ts
@@ -5,7 +5,7 @@ import { songlistToTracklist } from '@objects/Playlist';
 import noxPlayingList, { getCurrentTPQueue } from '@stores/playingList';
 import biliavideo from '@utils/mediafetch/biliavideo';
 import { randomChoice } from '@utils/Utils';
-import { NoxRepeatMode } from './enums/RepeatMode';
+import { NoxRepeatMode } from '@enums/RepeatMode';
 
 const { getState } = noxPlayingList;
