+++ src/enums/RepeatMode.ts
